requirements.txt
pandas==1.5.3
numpy==1.24.3
faker==18.13.0
matplotlib==3.7.2
seaborn==0.12.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
schedule==1.2.0
tqdm==4.66.1
python-dotenv==1.0.0
plotly==5.16.1
pyyaml==6.0.1